Handle duplicated values on imports with cartodb_id
When trying to import a shapefile with duplicated or invalid values on a column
named cartodb_id the importer will now rename the column to invalid_cartodb_id
and create a new cartodb_id column to be used as primary key.

CSV imports don't have this problem, ogr2ogr handles duplicates.