save info window state for each schema
When a query is applied and the info window fields
are changed according to the query they may be reset 
if the user changes the query (to do a temporal analysis 
query for example).

This code saves sue state so when the user come back 
to the desired query infowindow fields are restored.