Revoke privs before dropping org user. CDB avatars
Before trying to drop organization user schema revoke all privileges of
the user and the public user in all schemas and schema objects of the
database.

Implemented CartoDB avatars. If the user doesn't have a gravatar, try to
load a CartoDB avatar. If can't load the CDB avatar load the default
avatar