Don't try to initialize calendar that can't be initialized
The initializing of the datepicker plugin is done in a timeout in the
view’s constructor, so the view has time to be injected to the
document’s DOM, because the DatePicker plugin requires that.

In tests there’s a bunch of error output due to this, so make sure the
calendar is present in the document before trying to initialize it.