Use the real index of the layer to enable the interaction.
When some of the layers where hidden, layerView.getLayerNumberByIndex(i) was returning -1, and we were enabling the interaction of layer -1.  With this change, we are enabling the interaction of hidden layers if they have a infowindows. As an alternative, we could check if index == -1 here and prevent interaction from being enabled.