Implement common paged_searched behavior
Now applying to organization users and grantables, to have a common
behavior for how params (search+pagination in this case) are handled
for a collection.

To achieve this, I also had to move models to a browserify bundle
instead, although the models will continue to work as before (since
attached to global cdb.admin object). This change however enables
requires to be possible for models, as well as source maps!

Organisation model had to bundled in separate browserify bundle too
unfortunately , for non-browserify bundles that depends on it to work
though (due to it having a require).