Implement time-series as a separate widget
Reuses the histogram model, but need to translate widget definition’s
type to ‘histogram’ before creating the model for communication with
the tiler to work as expected. Due to this I also had to refactor the
model creation (followed same approach as for the widget view factory).
cc @alonsogarciapablo @javierarce @rochoa