Disallow changing type of an instantiated layer
Just changing attributes doesn’t make it work as expected, since the
old model is still used. Since this shouldn’t be allowed I’ve changed
the behavior to log an error if someone tries to do this.

Updated the affected specs accordingly.

Will also affect editor, to be handled in
https://github.com/CartoDB/cartodb/issues/6068