Handle chart resize internally
This DRY the code duplication as can be seen in the changes, by making
the chart to adapt to its parent container’s size in a consistent way.

Added a “resized” event to allow parent owner to update other things if
necessary (e.g. Torque’s time-marker view).

cc @javierarce