Implement dashboard layout view
Since the existing vis view expects a map element it’s wrapped in the
top-level creation, to not have to migrate all the existing code there
now

Separated the view factory setup in two instances, for the different
contexts that exist now (sidebar and below-map), that way the
definitions are simple and in-context.

Created a top-level collection for the widgets, since the
widgets-realted views have no need to know about layers.