Raise exception when CSV file with XLS format is detected
Sometimes, plain text CSV files with a XLS extension are uploaded
and they were being treated in this piece of code. It seems that
the in2csv process gets stuck with big (80K rows) CSV files but it works
smoothly with some other smaller files (3K rows).

This commit attempts to detect those false XLS files in order to
avoid resques to get stuck indefinitely. When these files are detected,
a Malformed XLS exception will be raised to the user.