Fixes #5131: https://github.com/CartoDB/cartodb/issues/5131
Map failing when cartocss is using a non-existing column. This is due a schema change after a successful sync,
dropping column/s. Included a verification for it, alerting the user through the
UI. Includes a new state for the Synchronization, modified, this state would
became the defacto when the schema has changed and the new sync contains less
columns than the previously synced table.