Check import URLs against a blacklist of IP addresses
Currently, the importer does not perform any kind of check when a
URL is requested to be imported. In production environments, this
blacklist can be used in order to block imports from specific IPs
to improve security measures.

This commit adds a new entry in the app_config that allows to specify
specific IP addresses or ranges of IP addresses from which
imports will not be allowed.