Check final URL after Typhoeus request
The original URL validator checks the initial URL that
is provided by the user, but this validation check could
be tricked in case that the checked URL performs a redirect.

Now, the redirect_count is checked after Typhoeus's first
request to know if the URL performs a redirect. In case it does,
the final URL (effective_url) is checked again against the
url_validator.

By checking the redirect_count value we avoid running the
validator in those cases which are already covered by the validation
performed by the import controller.

Style fixes

Fix tests