Push dataview validation and attrs definitions down to cartodb.js
Didn’t really belong here in the first place, this resolved the issue
of complicated logic to decide what attrs goes where.

DI `_.pick`s whatever it needs, in the same way as cartodb.js can
filter out what attrs are relevant to create the dataview.

This way, validation of the attrs here are _only_ for the widgets, and
dataviews should know how to validate themselves too. Errors are let to
bubble up in case of errors, so we (as devs) gets aware as early as
possible.

cc @alonsogarciapablo