Remove option of changing layer for a widget
After offline discussion we agreed this complicated things, one get to
select the layer when creating the widget, from there it’s just allowed
to modify the settings within that selection.

To “change layer” the user would have to delete the current widget and
create a new one, selecting the wanted layer.