Avoid arcgis urls limitation by arcgis webserver name
The ArcGIS Server endpoint URLs can look like http://<host>/<site>/rest/services/<folder>/<serviceName>/<serviceType>/
so this commit is removing the <site> limitation that was being imposed
with the VALID_ARCGIS_WEBSERVERS constant and looks for
"/rest/services" in the URLs instead. This opens the possibility of
ArcGIS server imports to all those services where the "site" name is not
the common or default ones.