Store analysis models as separately objects
In lieu of nesting them when serialized, which otherwise would lead to
duplicates nodes stored in DB when there are multiple heads to a
particular node.

This also makes the persisted models to match the editor ones, which
makes it trivial to restore those models on editor startup (due to a
full-page reload for example).