Change debug notice to identify non-downloadable gdrive files
People can share files in GDrive and select in the advanced sharing
options that the users with which these files are shared are not
authorized to download the files. This was causing CartoDB to crash, as
it was expecting a DownloadUrl or an ExportLink, and in this
"non-downloadable" scenario none of the options is available for the
item.

In this case, we just ignore the non-downloadable files and list the
rest of the files. In case we need to debug this in the future, the
non-downloadable debug message can be found at the notifier.