Static map url helper improvements
When an org user shared a map or dataset with another user inside the
org and this second user likes the visualization, the URL generated to
show a preview image for that visualization in the notification email
was using the subdomain of the second user, not the owner. Because of
that, the URL was returning a 404 due to that user hasn’t got that
static map.

The solution was creating a new method used for the same purpose (refer
the static map for the liked map/dataset) taking the visualization
owner’s username to create the URL correctly. Many thanks to @gfiorav
for his help.