Fix read/write table rename on datasets editor view
There were some missing methods/text keys from the move,
this puts them in place and adds an additional attempt to
pull the current table model from the TabPanel active pane
if other attempts fail. This case is valid when we viewing
a read/write table on the datasets editor view.