Infobox buttons improvements
Added new infobox parameter `position` needed to set the buttons, this
allows us to decide where we want the button to be (right or left).
With this modifications and the substitution in the templates of the
selectors `.js-primary/secondary` with `.js-mainAction/secondAction`
now we also can decide the action assigned to the button no matter the
position of it, while before it was directly attached to it.