Checked if c is defined.
If the user tries to generate a range from a column where all values are null, the domain in the
style generated doesn't have any element. When we are generating the categories colors, we tried to
unquote the color name before to apply the empty string fallback. This fixes that case.