Wait server response to add the model.
Not waiting caused a race condition and two creation requests were made. The first one was aborted
by the sync abort thing, but the model was already added to the collection. Then the style
definition model changed, causing the widgets to persist.

So two time series were created when this race condition met.