Codemirror mini mustache mode.
Added styles for mode support.
Include option to customize the number of keys to show autocomplete. Before was 3.
Include an option for triggers, besides keywords. This way whenever any trigger hits a match, the
autocomplete shows.
errorTemplate and tip are now optional for codemirror view.
Include options to generate subset of hints. Now we can choose to not show tableName or keywords for
instance. This way `tokens` became optional.