Check SAML before LDAP
This is needed to ensure CSRF is checked in all cases. In particular
there was a corner case if you enabled LDAP + SAML. In that case,
the CSRF hook was disabled by SAML, and then run manually during SAML
authentication checks. However, as LDAP ran first, this was effectively
skipped.