Chaining requests.
Until now, we was doing the requests in parallel. But this failed when the measurement changed and
the values for normalized for instance is not available for this measure, so basically all the
requests were wrong. In order to fix this issue, we need to do the request in sequence, trying to
sync the value set with the values returned by the api before doing the next request.