If measurements view has a value, return the value.
Our form controls are not ready to manage async data. Here we have to deal with fetch collections
when the component has received a value from the form model initialization. So, if there is an item
selected in the collection, return it. If not and there is a value, use this instead.