Don't listen region changes.
Listening region changes provokes memory leaks. Every region change, trigger a new form render so a
new model is created. But the former model is on the region change promise train, so it remains in
memory, but nobody it's listening to it though. So to react to a region changes, we check if
measurement is set on initialization.