Fix thumbnails generation (center, zoom, bounds)
Fix the thumbnail generation when a map is updated. When the named map
is updated, instead of using the center, zoom and bounds of the Map
model, use the state of the visualization. The Map models always remain
unchanged. This way the named map contains the right information to
render the correct viewport when requesting an image to the static map
API.