Decoupling back action and clean.
In editor-map-view we are listening changes in map-mode-model. The
problem is the stack-layout-view performs a clean call in the current
view. And the clean changes the map-mode-model mode. So, if being
editing or adding a feature, you click on edit widget link, this changes
the stack layout model position to point widget view. But the clean
changes the map mode, and the listener call stackLayoutModel.goBack,
trying to load the feature edition view, without a feature this time,
throwing a couple of errors.

So the solution implemented has been decoupling back action of cleaning
action. Back changes the map mode normally, so the stack layout goes
back without any issue. But clean method calls internally a new reset
method on map mode model, that updates silently the mode without
triggering any listener.