Applied sanitize labels layer.
When the collection is populated by the first time, so the map is
autofixed initially. But also, as a safety net, we also sanitize them
when the basemap is changed. In terms of performance there is a minor
impact using this function: ~24 miliseconds when the layer is destroyed
and ~0.05 miliseconds when the basemap changes.