Solution to wrong ACL (deleted user)
Here is the scenario we're trying to fix:

- some user A creates a table
- that user A grants permission to some other user B to that table
- the user B is deleted

Usually this is dealt with as part of the user deletion, but we found a
situation in which ACL (permissions table) is wrong because they
reference a deleted user. We don't know exactly how we arrived to that
situation. It is most likely a race condition deleting a user or migrating
a DB, but we don't have enough information to reconstruct the whole
story.

The trouble with that is that the web client is generating requests to
change permissions based on the wrong (referencing the deleted user)
ACL.

Once you have the data inconsistent in the DB you have 3 possible
approaches to get it fixed:

- fix it directly in the database
- filter the input of PUT /api/v1/perm, ignoring deleted entities
- filter the output (viz endopoint, permissions presenter)

Fixing directly in the database is OK as a one-time thing, but that
may cause some maintenance burden.

Returning an error if the input is wrong is OK. Filtering for wrong
inputs could fix it but I don't think it's the right approach, as some
other errors could be masked otherwise.

I think filtering the output from DB is the rigth approach. This way the
permissions can be set correctly by the user. They will get "automatically
fixed" the next time the user tries to modify the permissions.

And this is why the fix for it is in the presenter.