Export .carto.gpkg metadata
In the event that the imported file is a .carto.gpkg, this commit adds
functionality to recover the metadata from this file through a
metadata attr_reader. It employs geopkg_carto_metadata_util to do so.
This commit also organizes attr_reader at the beginning of the class
in hopes the trend will catch up.