Store import metadata in DataImport
This commit adds getters and setters to the DataImport model, to allow
for import metadata storage. Sequel model's limited capabilities to
deal with JSON has lead to overriding said getter/setters. Invalid
JSON will be ignored and return nil in getter. Setter requires a Hash
type to be passed in.