Fix autostyle category
According to the [turbo-carto-documentation](https://github.com/CartoDB/turbo-carto) we want "=" mapping for
category widgets.

In the past we were specifying no mapping for the ramps, but since our data was not numeric, turbocarto was smart
enough to create a category-ramp.

This commit fixes the ramp generator, passing a "=" mapping for the category case.