Fix for the issue of tmp dir with capital letters
Some might be wondering how this fixes it. The problem is that that
`.downcase` is downcasing the full path, including the directory and
the file name.

Later on, it tries to retrieve the file from a different path, and the
file is not there.

What sense does it have to downcase a filename? Actually no sense:
table name normalization is taken care of in other parts of the code,
and we can afford the original file names with or without caps.