Bolt and importer changes (#14736)
* Execute code block even if Bolt is locked

You probably are wondering why is the purpose of this. The main reason
to execute a "critical section" without acquiring the mutex (Bolt) is
because we only need that the lock has been acquired so we know for sure that Ghost Tables are not going to be executed.

In which cases are we going to need this? Mainly in the register part of
the importer where we're going to perform DDL operations that are going
to trigger ghost tables jobs and that could coincide at the same time
with other DDL operations leading into a race condition.

* Retry bolt block of code if needed

We've added the possibility to retry critical sections of code if Bolt
received an attempt to lock and that block of code is marked as
retriable.

Why are we doing this? Because we don't want to miss changes in the
user's schema because Bolt was already acquired.

For example, if some user is doing intensive DDL operations we want to
sync all the changes so to achieve that we need to inform Bolt that
we have other processes trying to execute Ghost tables at the same
time and instead of releasing the lock, we mark it asking for a retry
and at the end, we re-execute the same critical section again.