Adding `location` parameter to the import API
If the user uses Storage API and the dataset is located in different
location than "US" then they have to provide the `location` parameter.
With this parameter the connector creates a temp dataset in that
location if it does not exist in order to allow ODBC driver to create
temp tables.