Remove FROM unpackaged
This is hyper legacy. The purpose was to grab disparate objects out of
an extension within an extension. That predates even the
cartodb-postgresql extension and was a convenience for one or zero
developers.

The cartodb-postgresql extension make process generates "FROM
unpackage" upgrade paths just to the last version.

So removing that is simplifying our lives and getting rid of legacy.