Apply get_valid_column_name in tests properly
The affected tests were applying name sanitization to multiple columns
in a way that is, in principle, incorrect.
The get_valid_column_name is not designed to be passed the columns
incrementally; that would produce non-unique column names
(e.g. for [' ', '_'])
This modification has not practical effects because the inputs used
in the tests are immune to this problem, but we better do it properly,
so that future modifications to the tests or future tests copied from
those are correct.