Fix overview tests
The code checked that no error was reported with CartoDB::Logger
before. But syncs and imports use their own logger, so it didn't make
sense anyway.

Aside, I checked and the exceptions are all being captured and dealt
with and do not bubble up, so the errors because of the exceptions I
think are due to an unwanted interaction with the test framework.

Anyway, both tests and code are perfectly functional, so removing that
extra check.