Use changed attrs to check if the user is valid
When the user is being updated the method valid_user? uses the changed
attributes to check if the user is valid instead of checking the
persisted values which are outdated.

This prevent removing the api_keys from redis for a locked user when is
unlocked by changing its price plan. This happened because it checks if
the user is valid by reading the persisted values which are outdated.