Fix loading of CartoDB::PermissionPresenter dependency
The class CartoDB::PermissionPresenter is not following the
folder/module path conventions and the autoloader may not be able to
load it.

Seen in tests (it may depend on execution order):

```
Failures:

  1) Carto::Api::VisualizationsController index returns valid information for a user with one table
     Failure/Error: table1_visualization_hash = table1_visualization.to_hash(
     NameError:
       uninitialized constant Permission::PermissionPresenter
     # ./app/models/visualization/presenter.rb:32:in `to_poro'
     # ./app/models/visualization/member.rb:317:in `to_hash'
     # ./spec/requests/carto/api/visualizations_controller_spec_part_1.rb:317:in `block (3 levels) in <top (required)>'

Finished in 15.05 seconds
1 example, 1 failure

Failed examples:

rspec ./spec/requests/carto/api/visualizations_controller_spec_part_1.rb:314 # Carto::Api::VisualizationsController index returns valid information for a user with one table
```

this makes it robust, but the proper fix is to follow the conventions.