Update pubsub to 2.3.0
Update pubsub to ~> 2.3 and some other dependencies. In particular I
had to update gapic manually in order to solve a conflict, by issuing
the following:

    bundle update gapic-common

A more aggressive update can be performed with `bundle update` (no
extra gem) but did not want to run the risk (it can be done in a
latter PR).