Do not attempt to check username/email against central in on-prem
This happens when a new SAML or LDAP user signs up and it is not yet
created in cartodb. It results in a sync call to Central HTTP API.

This avoids such call when the Central api is not enabled (notably in
on-premise installations).

Also this fixes tests that otherwise wait forever on an `until`.